chore(acp): fix module paths after rebase onto restructured main
The upstream src reorg moved settings/config_service into crate::config,
mcp_e2e_tests into crate::testing::mcp_e2e, etc. Update the ACP feature's
imports and widen the mcp_e2e fixture module to pub(crate) so the ACP
e2e suite can reuse the real echo-server fixture.
